# Conflicts:
#	src/main/java/com/project/JewelryMS/controller/OrderController.java
#	src/main/java/com/project/JewelryMS/controller/StaffAccountController.java
#	src/main/java/com/project/JewelryMS/service/ProductBuyService.java

+++ src/main/java/com/project/JewelryMS/repository/ShiftRepository.java
@@ -14,7 +14,7 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface ShiftRepository extends JpaRepository<Shift, Integer> {
+public interface ShiftRepository extends JpaRepository<Shift, Long> {
 
     // Find shifts by start time
     @Query("SELECT s FROM Shift s WHERE s.startTime = :startTime")
